memoryviewer, and breakpoints, fixed rendering of 'negative' instructions

+++ src/main/java/de/thkoeln/ra/team3/risc_thk_simulator/guiTemplates/InstructionEntry.java
@@ -22,7 +22,7 @@ public class InstructionEntry {
 		} memAddr = "0x"+memAddr;
 		breakpoint = new CheckBox(memAddr);
 		
-		String irHex = Long.toHexString(instruction.getIr());
+		String irHex = Integer.toHexString(instruction.getIr());
 		while(irHex.length() != 8) {
 			irHex = "0"+irHex;
 		}
